Move optimization loop info functoin
